58-image-classifier/requirements.txt
streamlit>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
requests>=2.31.0
